+++ src/233_23_CS_Camera_ThTick_Podium.c
@@ -122,9 +122,9 @@ void CS_Camera_ThTick_Podium(struct Thread *th)
         && (sdata->ptrActiveMenu == NULL)
     )
     {
-        int auto_skip_podium = false; // default
+        unsigned short auto_skip_podium = false; // default
         int db_result = DB_VALUE_NOTFOUND;
-        int db_ret = database_fetch(
+        unsigned short db_ret = database_fetch(
             (DB_PREFIX_SETTINGS | SETTING_QOL_SKIP_PODIUM) << 16,
             &db_result
         );
@@ -195,9 +195,9 @@ void CS_Camera_ThTick_Podium(struct Thread *th)
 
                 // If the gem color is == 5 then we have a multiworld item gem.
                 // In this case, or if we always skip the hints here: no hint
-                int skip_mask_congrats = false; // default
+                unsigned short skip_mask_congrats = false; // default
                 int db_result = DB_VALUE_NOTFOUND;
-                int db_ret = database_fetch(
+                unsigned short db_ret = database_fetch(
                     (DB_PREFIX_SETTINGS | SETTING_QOL_SKIP_CONGRATS) << 16,
                     &db_result
                 );
