+++ src/AH_Garage_ThTick.c
@@ -160,9 +160,9 @@ LAB_800aeb6c:
     }
     else // If you're not in Gemstone Valley
     {
-        int garage_requirement_logic = GARAGE_OPENING_VANILLA_WARPPADS; // default
+        unsigned short garage_requirement_logic = GARAGE_OPENING_VANILLA_WARPPADS; // default
         int db_fetch_result = DB_VALUE_NOTFOUND;
-        int db_ret = database_fetch(
+        unsigned short db_ret = database_fetch(
             (DB_PREFIX_SETTINGS | SETTING_BOSS_GARAGE_OPENING) << 16,
             &db_fetch_result
         );
@@ -206,7 +206,7 @@ LAB_800aeb6c:
                 {
                     // Adjust check to look at the warp pads in the current hub
                     // compensating for any warp pad randomization
-                    int randomized_LevelID = database_fetch(
+                    unsigned short randomized_LevelID = database_fetch(
                         (DB_PREFIX_LEVELIDS | levelID_to_check) << 16,
                         &db_fetch_result
                     );
@@ -333,9 +333,9 @@ LAB_800aede8:
 
         if (levelID == GEM_STONE_VALLEY)
         {
-            int oxide_relics_required = RELIC_SAPPHIRE; // default
+            unsigned short oxide_relics_required = RELIC_SAPPHIRE; // default
             int db_result = DB_VALUE_OK;
-            int db_ret = database_fetch(
+            unsigned short db_ret = database_fetch(
                 (DB_PREFIX_SETTINGS | SETTING_OXIDE_REQUIRED_RELICS) << 16,
                 &db_result
             );
