+++ src/AH_Map_HubItems.c
@@ -143,9 +143,9 @@ void AH_Map_HubItems(void* hubPtrs, short *param_2)
                         // must be == 3, for Boss Garage
                         int base = levelID - N_SANITY_BEACH;
 
-                        int garage_requirement_logic = GARAGE_OPENING_VANILLA_WARPPADS; // default
+                        unsigned short garage_requirement_logic = GARAGE_OPENING_VANILLA_WARPPADS; // default
                         int db_fetch_result = DB_VALUE_NOTFOUND;
-                        int db_ret = database_fetch(
+                        unsigned short db_ret = database_fetch(
                             (DB_PREFIX_SETTINGS | SETTING_BOSS_GARAGE_OPENING) << 16,
                             &db_fetch_result
                         );
@@ -188,7 +188,7 @@ void AH_Map_HubItems(void* hubPtrs, short *param_2)
                                 {
                                     // Adjust check to look at the warp pads in the current hub
                                     // compensating for any warp pad randomization
-                                    int randomized_LevelID = database_fetch(
+                                    unsigned short randomized_LevelID = database_fetch(
                                         (DB_PREFIX_LEVELIDS | levelID_to_check) << 16,
                                         &db_fetch_result
                                     );
